Relocate Data Helper Folder
