+++ data/src/main/java/com/hungho/data/local/database/entity/UserRemoteKeyEntity.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "user_remote_keys")
-data class UserRemoteKeyEntity(
+internal data class UserRemoteKeyEntity(
     @PrimaryKey val userId: Int,
     val prevKey: Int?,
     val nextKey: Int?
